Practiced some JAVA imp Concepts
This File contains a Randomised elements in  array and using the Arrays inbuilt method os Sorting and reversing is Used.Operations such as Fill and Sorting again is applied.Also Implemented Characters inbuilt methods of .isDigit and .isLetter are used to make verification easier.
Also implemented comcept of finally and implemeted.
Finding Area is a java program of Inheritance and Method Overriding Combined.